Time Slider: Move Windows-specific logic to partial class file
